Re-Organizing: SystemUi Settings WIP Pth 1
Out of my missing months, i just realized, we don't really need to put those ugly sliders in this tool where we can just make it simple.

This commit redefines the few later changes of the slider back to stock. I'm removing the slider, and improving on the UX design, so this is somehow the new era for Carbon AOSP.

Haven't debug and tested yet. Not yet assigned a designer for the icons.

Signed-off-by: Daniel MM <daxx4182@gmail.com>